chore: Add common dir to prompt for script

+++ scripts/components/generate-component.ts
@@ -12,17 +12,11 @@ const [name] = process.argv.slice(2);
 
 if (!name) throw new Error('Please provide a component name.');
 
-// Create the directory to hold new React component files.
-const directory = `./src/components/${name}`;
-if (fs.existsSync(directory))
-  throw new Error('A component with that name already exists.');
-fs.mkdirSync(directory);
-
 /**
  * Create necessary React component files like
  * Component, SCSS and Storybook stories files.
  */
-async function createComponentFiles(): Promise<void> {
+async function createComponent(): Promise<void> {
   prompt.start();
   const schema = {
     properties: {
@@ -40,21 +34,41 @@ async function createComponentFiles(): Promise<void> {
         warning: 'You must choose Yes or No!',
         default: 'no',
       },
+      isCommon: {
+        name: 'isCommon',
+        message: 'Will the component be shared?',
+        validator: /y[es]*|n[o]?/,
+        warning: 'You must choose Yes or No!',
+        default: 'no',
+      },
     },
   };
   await prompt.get(schema, (error, results) => {
+    // Set the root of the component based on the result from the prompt.
+    let root = 'components';
+    if (results.isCommon === 'yes' || results.isCommon === 'y') {
+      root = 'common';
+    }
+    // Create the directory to hold new React component files.
+    const directory = `./src/${root}/${name}`;
+    if (fs.existsSync(directory))
+      throw new Error('A component with that name already exists.');
+    fs.mkdirSync(directory);
+
     // Create the actual .tsx React component file.
     fs.writeFile(
       `${directory}/${name}.tsx`,
       component(
         name,
-        results.hasInterface === 'yes' || 'y' ? 'withInterface' : ''
+        results.hasInterface === 'yes' || results.hasInterface === 'y'
+          ? 'withInterface'
+          : ''
       ),
       writeErrorHandler
     );
     // Create the new component's stylesheet.
     fs.writeFile(`${directory}/${name}.scss`, '', writeErrorHandler);
-    if (results.hasStory === 'yes' || 'y') {
+    if (results.hasStory === 'yes' || results.hasStory === 'y') {
       // Create the new component's Storybook story.
       fs.writeFile(
         `${directory}/${name}.stories.tsx`,
@@ -65,4 +79,4 @@ async function createComponentFiles(): Promise<void> {
   });
 }
 
-createComponentFiles();
+createComponent();
